Moved the methods to Acq Class
