+++ SemiconductorTestLibrary.Extensions/source/InstrumentAbstraction/Scope/Acquisition.cs
@@ -1,4 +1,8 @@
-﻿using NationalInstruments.ModularInstruments.NIScope;
+﻿using System.Linq;
+
+using NationalInstruments.ModularInstruments.NIScope;
+using NationalInstruments.SemiconductorTestLibrary.DataAbstraction;
+
 using static NationalInstruments.SemiconductorTestLibrary.Common.ParallelExecution;
 
 namespace NationalInstruments.SemiconductorTestLibrary.InstrumentAbstraction.Scope
@@ -44,6 +48,40 @@ namespace NationalInstruments.SemiconductorTestLibrary.InstrumentAbstraction.Sco
             return sessionsBundle.DoAndReturnPerInstrumentPerChannelResults(sessionInfo =>
             {
                 return sessionInfo.Session.Acquisition.SampleRate;
+            });
+        }
+
+        /// <summary>
+        /// Reads waveform data from all channels in the bundle.
+        /// </summary>
+        /// <param name="sessionsBundle">The <see cref="ScopeSessionsBundle"/> object.</param>
+        /// <param name="timeout">The maximum time to wait for the data.</param>
+        /// <param name="recordLength">The number of samples to read.</param>
+        /// <returns>An array of waveforms, where each element corresponds to one channel in the bundle.</returns>
+        public static PinSiteData<AnalogWaveform<double>> ReadWaveform(this ScopeSessionsBundle sessionsBundle, PrecisionTimeSpan timeout, long recordLength)
+        {
+            return sessionsBundle
+                .DoAndReturnPerSitePerPinResults((sessionInfo, sitePinInfo) =>
+                {
+                    return sessionInfo.Session.Channels[sitePinInfo.IndividualChannelString]
+                        .Measurement
+                        .Read(timeout, recordLength, null)
+                        .First();
+                });
+        }
+
+        /// <summary>
+        /// Fetches waveform data from the specified channel without initiating a new acquisition.
+        /// </summary>
+        /// <param name="sessionsBundle">The <see cref="ScopeSessionsBundle"/> object.</param>
+        /// <param name="timeout">The maximum time to wait for the data.</param>
+        /// <param name="pointsToFetch">The number of points to fetch. Use -1 to fetch all available points.</param>
+        /// <returns>An array of waveforms, where each element corresponds to one channel in the bundle.</returns>
+        public static PinSiteData<AnalogWaveform<double>> FetchWaveform(this ScopeSessionsBundle sessionsBundle, PrecisionTimeSpan timeout, int pointsToFetch)
+        {
+            return sessionsBundle.DoAndReturnPerSitePerPinResults((sessionInfo, sitePinInfo) =>
+            {
+                return sessionInfo.Session.Channels[sitePinInfo.IndividualChannelString].Measurement.FetchDouble(timeout, pointsToFetch, null).First();
             });
         }
     }
